+++ packages/tempus/test/BDF2/Tempus_BDF2_Pseudotransient_FSA.cpp
@@ -1,0 +1,23 @@
+// @HEADER
+// ****************************************************************************
+//                Tempus: Copyright (2017) Sandia Corporation
+//
+// Distributed under BSD 3-clause license (See accompanying file Copyright.txt)
+// ****************************************************************************
+// @HEADER
+
+#include "Tempus_BDF2_FSA.hpp"
+
+namespace Tempus_Test {
+
+TEUCHOS_UNIT_TEST(BDF2, SteadyQuadratic_PseudoTransient_FSA)
+{
+  test_pseudotransient_fsa(false, out, success);
+}
+
+TEUCHOS_UNIT_TEST(BDF2, SteadyQuadratic_PseudoTransient_FSA_Tangent)
+{
+  test_pseudotransient_fsa(true, out, success);
+}
+
+} // namespace Tempus_Test
