+++ packages/tempus/test/BDF2/Tempus_BDF2_SinCos_SA.xml
@@ -1,6 +1,14 @@
-<ParameterList name="BackwardEuler_SteadyQuadratic">
-  <ParameterList name="SteadyQuadraticModel">
-    <Parameter name="Coeff b" type="double" value="1.0"/>
+<ParameterList name="BDF2_SinCos">
+  <ParameterList name="SinCosModel">
+    <Parameter name="Accept model parameters"    type="bool" value="true"/>
+    <Parameter name="Provide nominal values"     type="bool" value="true"/>
+    <Parameter name="Coeff a" type="double" value="0.0"/>
+    <Parameter name="Coeff f" type="double" value="1.0"/>
+    <Parameter name="Coeff L" type="double" value="1.0"/>
+    <Parameter name="IC x0"   type="double" value="0.0"/>
+    <Parameter name="IC x1"   type="double" value="1.0"/>
+    <Parameter name="IC t0"   type="double" value="0.0"/>
+    <Parameter name="Number of Time Step Sizes" type="int" value = "7"/>
   </ParameterList>
   <ParameterList name="Tempus">
     <Parameter name="Integrator Name" type="string" value="Default Integrator"/>
@@ -16,12 +24,12 @@
       </ParameterList>
       <ParameterList name="Time Step Control">
         <Parameter name="Initial Time"           type="double" value="0.0"/>
-        <Parameter name="Final Time"             type="double" value="20.0"/>
+        <Parameter name="Final Time"             type="double" value="1.0"/>
         <Parameter name="Initial Time Index"     type="int"    value="0"/>
         <Parameter name="Final Time Index"       type="int"    value="10000"/>
         <Parameter name="Minimum Time Step"      type="double" value="0.00001"/>
-        <Parameter name="Initial Time Step"      type="double" value="1.0"/>
-        <Parameter name="Maximum Time Step"      type="double" value="1.0"/>
+        <Parameter name="Initial Time Step"      type="double" value="0.1"/>
+        <Parameter name="Maximum Time Step"      type="double" value="0.1"/>
         <Parameter name="Minimum Order"          type="int"    value="1"/>
         <Parameter name="Initial Order"          type="int"    value="1"/>
         <Parameter name="Maximum Order"          type="int"    value="2"/>
