+++ packages/tempus/test/BDF2/Tempus_BDF2_Staggered_FSA.cpp
@@ -1,0 +1,23 @@
+// @HEADER
+// ****************************************************************************
+//                Tempus: Copyright (2017) Sandia Corporation
+//
+// Distributed under BSD 3-clause license (See accompanying file Copyright.txt)
+// ****************************************************************************
+// @HEADER
+
+#include "Tempus_BDF2_FSA.hpp"
+
+namespace Tempus_Test {
+
+TEUCHOS_UNIT_TEST(BDF2, SinCos_Staggered_FSA)
+{
+  test_sincos_fsa(false, false, out, success);
+}
+
+TEUCHOS_UNIT_TEST(BDF2, SinCos_Staggered_FSA_Tangent)
+{
+  test_sincos_fsa(false, true, out, success);
+}
+
+} // namespace Tempus_Test
