+++ packages/tempus/test/BackwardEuler/CMakeLists.txt
@@ -9,8 +9,29 @@ TRIBITS_ADD_EXECUTABLE_AND_TEST(
   )
 
 TRIBITS_ADD_EXECUTABLE_AND_TEST(
-  BackwardEulerSensitivity
-  SOURCES Tempus_BackwardEulerSensitivityTest.cpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  BackwardEuler_Combined_FSA
+  SOURCES Tempus_BackwardEuler_Combined_FSA.cpp Tempus_BackwardEuler_FSA.hpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  TESTONLYLIBS tempus_test_models
+  NUM_MPI_PROCS 1
+  )
+
+TRIBITS_ADD_EXECUTABLE_AND_TEST(
+  BackwardEuler_Staggered_FSA
+  SOURCES Tempus_BackwardEuler_Staggered_FSA.cpp Tempus_BackwardEuler_FSA.hpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  TESTONLYLIBS tempus_test_models
+  NUM_MPI_PROCS 1
+  )
+
+TRIBITS_ADD_EXECUTABLE_AND_TEST(
+  BackwardEuler_Pseudotransient_FSA
+  SOURCES Tempus_BackwardEuler_Pseudotransient_FSA.cpp Tempus_BackwardEuler_FSA.hpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  TESTONLYLIBS tempus_test_models
+  NUM_MPI_PROCS 1
+  )
+
+TRIBITS_ADD_EXECUTABLE_AND_TEST(
+  BackwardEuler_ASA
+  SOURCES Tempus_BackwardEuler_ASA.cpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
   TESTONLYLIBS tempus_test_models
   NUM_MPI_PROCS 1
   )
