+++ packages/tempus/test/DIRK/CMakeLists.txt
@@ -9,8 +9,29 @@ TRIBITS_ADD_EXECUTABLE_AND_TEST(
   )
 
 TRIBITS_ADD_EXECUTABLE_AND_TEST(
-  DIRKSensitivity
-  SOURCES Tempus_DIRKSensitivityTest.cpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  DIRK_Combined_FSA
+  SOURCES Tempus_DIRK_Combined_FSA.cpp Tempus_DIRK_FSA.hpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  TESTONLYLIBS tempus_test_models
+  NUM_MPI_PROCS 1
+  )
+
+TRIBITS_ADD_EXECUTABLE_AND_TEST(
+  DIRK_Staggered_FSA
+  SOURCES Tempus_DIRK_Staggered_FSA.cpp Tempus_DIRK_FSA.hpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  TESTONLYLIBS tempus_test_models
+  NUM_MPI_PROCS 1
+  )
+
+TRIBITS_ADD_EXECUTABLE_AND_TEST(
+  DIRK_Pseudotransient_FSA
+  SOURCES Tempus_DIRK_Pseudotransient_FSA.cpp Tempus_DIRK_FSA.hpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  TESTONLYLIBS tempus_test_models
+  NUM_MPI_PROCS 1
+  )
+
+TRIBITS_ADD_EXECUTABLE_AND_TEST(
+  DIRK_ASA
+  SOURCES Tempus_DIRK_ASA.cpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
   TESTONLYLIBS tempus_test_models
   NUM_MPI_PROCS 1
   )
