+++ packages/tempus/test/ExplicitRK/CMakeLists.txt
@@ -9,8 +9,29 @@ TRIBITS_ADD_EXECUTABLE_AND_TEST(
   )
 
 TRIBITS_ADD_EXECUTABLE_AND_TEST(
-  ExplicitRKSensitivity
-  SOURCES Tempus_ExplicitRKSensitivityTest.cpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  ExplicitRK_Combined_FSA
+  SOURCES Tempus_ExplicitRK_Combined_FSA.cpp Tempus_ExplicitRK_FSA.hpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  TESTONLYLIBS tempus_test_models
+  NUM_MPI_PROCS 1
+  )
+
+TRIBITS_ADD_EXECUTABLE_AND_TEST(
+  ExplicitRK_Staggered_FSA
+  SOURCES Tempus_ExplicitRK_Staggered_FSA.cpp Tempus_ExplicitRK_FSA.hpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  TESTONLYLIBS tempus_test_models
+  NUM_MPI_PROCS 1
+  )
+
+TRIBITS_ADD_EXECUTABLE_AND_TEST(
+  ExplicitRK_Pseudotransient_FSA
+  SOURCES Tempus_ExplicitRK_Pseudotransient_FSA.cpp Tempus_ExplicitRK_FSA.hpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
+  TESTONLYLIBS tempus_test_models
+  NUM_MPI_PROCS 1
+  )
+
+TRIBITS_ADD_EXECUTABLE_AND_TEST(
+  ExplicitRK_ASA
+  SOURCES Tempus_ExplicitRK_ASA.cpp ${TEUCHOS_STD_UNIT_TEST_MAIN}
   TESTONLYLIBS tempus_test_models
   NUM_MPI_PROCS 1
   )
